refactor: move SellNativeWarningBanner to tradeWidgetAddons
