+++ apps/cowswap-frontend/src/modules/tradeWidgetAddons/index.ts
@@ -4,5 +4,6 @@ export { SettingsTab } from './containers/SettingsTab'
 export { HighFeeWarning } from './containers/HighFeeWarning'
 export { BundleTxWrapBanner } from './containers/BundleTxWrapBanner'
 export { MetamaskTransactionWarning } from './containers/MetamaskTransactionWarning'
+export { SellNativeWarningBanner } from './containers/SellNativeWarningBanner'
 export { useHighFeeWarning } from './containers/HighFeeWarning/hooks/useHighFeeWarning'
 export { NetworkCostsTooltipSuffix } from './pure/NetworkCostsTooltipSuffix'
