+++ apps/cowswap-frontend/src/modules/twap/containers/TwapFormWarnings/index.tsx
@@ -3,11 +3,11 @@ import { useCallback } from 'react'
 
 import { useIsSafeViaWc, useWalletInfo } from '@cowprotocol/wallet'
 
-import { SellNativeWarningBanner } from 'modules/trade/containers/SellNativeWarningBanner'
 import { useTradeRouteContext } from 'modules/trade/hooks/useTradeRouteContext'
 import { useGetTradeFormValidation } from 'modules/tradeFormValidation'
 import { TradeFormValidation } from 'modules/tradeFormValidation/types'
 import { useTradeQuoteFeeFiatAmount } from 'modules/tradeQuote'
+import { SellNativeWarningBanner } from 'modules/tradeWidgetAddons'
 
 import {
   FallbackHandlerWarning,
